requirements.txt
Flask==1.0.2
numpy==1.16.2
Pillow==5.4.1
